# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NearbyPubsRequest {
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude: number;
}

export const AccountDeletionResponseValue = {
  ok: true,
} as const;
export type AccountDeletionResponse = typeof AccountDeletionResponseValue;

export type PintWarActivityEventType = typeof PintWarActivityEventType[keyof typeof PintWarActivityEventType];


export const PintWarActivityEventType = {
  pint_logged: 'pint_logged',
  pub_review: 'pub_review',
  legacy_pub_bonus: 'legacy_pub_bonus',
} as const;

export interface PintWarActivityEvent {
  /** @minLength 1 */
  id: string;
  type: PintWarActivityEventType;
  userId: string;
  /** @minLength 1 */
  playerName: string;
  occurredAt: string;
  /** @nullable */
  pubName: string | null;
  /** @nullable */
  scoreImpact: number | null;
  historicalScore: boolean;
  /** @nullable */
  photoPintLogId: string | null;
}

export interface PintWarActivityFeed {
  /** @maxItems 20 */
  events: PintWarActivityEvent[];
}

export interface Coordinates {
  latitude: number;
  longitude: number;
}

export type NearbyPubProvider = typeof NearbyPubProvider[keyof typeof NearbyPubProvider];


export const NearbyPubProvider = {
  google_places: 'google_places',
} as const;

export interface NearbyPub {
  provider: NearbyPubProvider;
  /** @minLength 1 */
  placeId: string;
  /** @minLength 1 */
  name: string;
  address: string;
  /** @minimum 0 */
  distanceMeters: number;
  coordinates: Coordinates;
}

export interface NearbyPubSearch {
  pubs: NearbyPub[];
  providerConfigured: boolean;
}

export interface ApiError {
  message: string;
}