# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: pubs
    description: Pub identification operations
  - name: account
    description: Account management operations
  - name: pint-wars
    description: Member-only Pint War activity and proof photos
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /pubs/nearby:
    post:
      operationId: findNearbyPubs
      tags: [pubs]
      summary: Find nearby pubs
      description: Returns Google Places pubs near an authenticated player's current location.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NearbyPubsRequest"
      responses:
        "200":
          description: Nearby pubs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearbyPubSearch"
        "400":
          description: Invalid coordinates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid Supabase session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Too many nearby-pub searches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Google Places request failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Places provider is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /account/delete:
    post:
      operationId: deleteMyAccount
      tags: [account]
      summary: Delete the authenticated account
      description: De-identifies account data and removes pint-proof storage objects while preserving historical league and score records.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account deletion and cleanup completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDeletionResponse"
        "400":
          description: Deletion is blocked by an active hosted Pint War or the account profile is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid Supabase session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Account data or storage cleanup could not be completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Account deletion is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /pint-proofs/leagues/{leagueId}/activity:
    get:
      operationId: getPintWarActivity
      tags: [pint-wars]
      summary: Get recent Pint War activity
      description: Returns a bounded, newest-first activity feed for an authenticated member of the Pint War.
      security:
        - bearerAuth: []
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PintWarActivityFeed"
        "400":
          description: Invalid Pint War identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid Supabase session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: The authenticated user is not a member of this Pint War
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Activity could not be loaded or membership could not be verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Activity service is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /pint-proofs/leagues/{leagueId}/photos/{pintLogId}:
    get:
      operationId: getPintWarActivityPhoto
      tags: [pint-wars]
      summary: Get a private Pint War proof photo
      description: Streams a proof photo only after verifying the caller is a member of the Pint War and the pint log belongs to that war.
      security:
        - bearerAuth: []
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pintLogId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Private proof image bytes
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/heic:
              schema:
                type: string
                format: binary
            image/heif:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid Pint War or pint log identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid Supabase session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: The authenticated user is not a member of this Pint War
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Pint log or proof image is missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "415":
          description: Unsupported proof image format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Proof image could not be loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Proof photo service is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    NearbyPubsRequest:
      type: object
      additionalProperties: false
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
      required:
        - latitude
        - longitude
    AccountDeletionResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
      required:
        - ok
    PintWarActivityFeed:
      type: object
      additionalProperties: false
      properties:
        events:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/PintWarActivityEvent"
      required:
        - events
    PintWarActivityEvent:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
        type:
          type: string
          enum: [pint_logged, pub_review, legacy_pub_bonus]
        userId:
          type: string
          format: uuid
        playerName:
          type: string
          minLength: 1
        occurredAt:
          type: string
          format: date-time
        pubName:
          type: [string, "null"]
        scoreImpact:
          type: [integer, "null"]
        historicalScore:
          type: boolean
        photoPintLogId:
          type: [string, "null"]
          format: uuid
      required:
        - id
        - type
        - userId
        - playerName
        - occurredAt
        - pubName
        - scoreImpact
        - historicalScore
        - photoPintLogId
    Coordinates:
      type: object
      additionalProperties: false
      properties:
        latitude:
          type: number
        longitude:
          type: number
      required:
        - latitude
        - longitude
    NearbyPub:
      type: object
      additionalProperties: false
      properties:
        provider:
          type: string
          enum: [google_places]
        placeId:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        address:
          type: string
        distanceMeters:
          type: number
          minimum: 0
        coordinates:
          $ref: "#/components/schemas/Coordinates"
      required:
        - provider
        - placeId
        - name
        - address
        - distanceMeters
        - coordinates
    NearbyPubSearch:
      type: object
      additionalProperties: false
      properties:
        pubs:
          type: array
          items:
            $ref: "#/components/schemas/NearbyPub"
        providerConfigured:
          type: boolean
      required:
        - pubs
        - providerConfigured
    ApiError:
      type: object
      additionalProperties: false
      properties:
        message:
          type: string
      required:
        - message