# lib/api-zod/src/generated/types/pintWarActivityEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PintWarActivityEventType } from './pintWarActivityEventType';

export interface PintWarActivityEvent {
  /** @minLength 1 */
  id: string;
  type: PintWarActivityEventType;
  userId: string;
  /** @minLength 1 */
  playerName: string;
  occurredAt: Date;
  /** @nullable */
  pubName: string | null;
  /** @nullable */
  scoreImpact: number | null;
  historicalScore: boolean;
  /** @nullable */
  photoPintLogId: string | null;
}