# lib/api-zod/src/generated/types/pintWarActivityFeed.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PintWarActivityEvent } from './pintWarActivityEvent';

export interface PintWarActivityFeed {
  /** @maxItems 20 */
  events: PintWarActivityEvent[];
}